Move fixtures and resource back to root of tests
Unit and integration may share fixtures and other resources. Move those directories from the unit directory back to the root of the tests directory. Refactor tests on where to look for these resources.
