+++ st2client/tests/fixtures/execution_get_detail.txt
