+++ st2client/tests/fixtures/execution_result_has_carriage_return.txt
