+++ st2client/tests/fixtures/loader.py
