+++ st2common/tests/resources/loadableplugin/manifest.yaml
