+++ st2common/tests/resources/loadableplugin/plugin/sampleplugin.py
