+++ st2common/tests/resources/loadableplugin/plugin/sampleplugin2.py
