+++ st2common/tests/resources/loadableplugin/plugin/sampleplugin3.py
