+++ st2common/tests/resources/loadableplugin/plugin/standaloneplugin.py
