+++ st2common/tests/resources/logging.conf
