+++ st2common/tests/resources/packs/pack1/rules/.gitignore
