+++ st2common/tests/resources/packs/pack1/sensors/.gitignore
