+++ st2common/tests/resources/packs/pack2/.gitignore
