+++ st2common/tests/services/test_action.py
