+++ st2reactor/tests/resources/sample_sensor_good_config.json
