+++ st2reactor/tests/resources/test_sensor.py
